refactor(formats): move read_* table functions into their format crates
Each file-format `read_*` table function now lives in its own `beacon-arrow-*`
crate (as `datafusion::table_function`, re-exported as `Read*Func`) instead of
all living in `beacon-functions`:

  read_parquet  -> beacon-arrow-parquet      read_netcdf -> beacon-arrow-netcdf
  read_csv      -> beacon-arrow-csv          read_atlas  -> beacon-arrow-atlas
  read_arrow    -> beacon-arrow-ipc          read_bbf    -> beacon-arrow-bbf
  read_geoparquet -> beacon-arrow-geoparquet read_tiff   -> beacon-arrow-tiff
  read_odv_ascii  -> beacon-arrow-odv        read_zarr   -> beacon-arrow-zarr

To avoid a dependency cycle (beacon-functions depends on the format crates), the
shared `BeaconTableFunctionImpl` trait and the `parse_glob_paths_arg` helper move
to `beacon-common` (which every format crate already depends on); beacon-functions
re-exports them for the cross-format functions.

`beacon-functions::register_table_functions` now constructs each per-format
function from its crate; the cross-format `read_schema` and `list_datasets` stay
in beacon-functions. Format crates that lacked them gained `tokio`/`tracing`
library deps.
