+++ beacon-common/src/lib.rs
@@ -5,5 +5,6 @@ pub mod listing_url;
 pub mod schema_table_provider;
 pub mod super_table;
 pub mod super_typing;
+pub mod table_function;
 
 pub use error::CommonError;
