+++ beacon-file-formats/beacon-arrow-geoparquet/src/datafusion/mod.rs
@@ -551,3 +551,6 @@ mod tests {
         }
     }
 }
+
+pub mod table_function;
+pub use table_function::ReadGeoParquetFunc;
