+++ beacon-file-formats/beacon-arrow-ipc/Cargo.toml
@@ -4,6 +4,8 @@ version = "1.7.3"
 edition = "2024"
 
 [dependencies]
+tracing = { workspace = true }
+tokio = { workspace = true }
 arrow = { workspace = true }
 datafusion = { workspace = true }
 object_store = { workspace = true }
