+++ beacon-file-formats/beacon-arrow-netcdf/Cargo.toml
@@ -11,6 +11,7 @@ arrow = { workspace = true }
 arrow-schema = { workspace = true }
 anyhow = { workspace=true }
 tempfile = { workspace=true}
+tokio = { workspace=true }
 futures = { workspace=true }
 indexmap = { workspace=true }
 thiserror = { workspace=true }
