+++ beacon-file-formats/beacon-arrow-odv/src/datafusion/table_function.rs
@@ -1,55 +1,56 @@
-use std::{fmt::Debug, sync::Arc};
+use std::sync::Arc;
 
 use arrow::datatypes::{DataType, Field};
 use beacon_common::{listing_url::parse_listing_table_url, super_table::SuperListingTable};
-use beacon_arrow_zarr::datafusion::ZarrFormat;
+use crate::datafusion::OdvFormat;
 use datafusion::{
-    catalog::TableFunctionImpl, execution::object_store::ObjectStoreUrl, prelude::SessionContext,
+    catalog::TableFunctionImpl,
+    execution::object_store::ObjectStoreUrl,
+    prelude::{Expr, SessionContext},
 };
 
-use crate::file_formats::BeaconTableFunctionImpl;
+use beacon_common::table_function::BeaconTableFunctionImpl;
 
-pub struct ReadZarrFunc {
-    // Session Reference
+pub struct ReadOdvAsciiFunc {
     runtime_handle: tokio::runtime::Handle,
     session_ctx: Arc<SessionContext>,
     data_object_store_url: ObjectStoreUrl,
 }
 
-impl ReadZarrFunc {
+impl ReadOdvAsciiFunc {
     pub fn new(
         runtime_handle: tokio::runtime::Handle,
-        session: Arc<SessionContext>,
+        session_ctx: Arc<SessionContext>,
         data_object_store_url: ObjectStoreUrl,
     ) -> Self {
         Self {
             runtime_handle,
-            session_ctx: session,
+            session_ctx,
             data_object_store_url,
         }
     }
 }
 
-impl Debug for ReadZarrFunc {
+impl std::fmt::Debug for ReadOdvAsciiFunc {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "ReadZarrFunc")
+        write!(f, "ReadOdvAsciiFunc")
     }
 }
 
-impl BeaconTableFunctionImpl for ReadZarrFunc {
+impl BeaconTableFunctionImpl for ReadOdvAsciiFunc {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
 
-    fn description(&self) -> Option<String> {
-        Some("Reads Zarr files from specified glob paths.".to_string())
-    }
-
     fn name(&self) -> String {
-        "read_zarr".to_string()
+        "read_odv_ascii".to_string()
     }
 
-    fn arguments(&self) -> Option<Vec<arrow::datatypes::Field>> {
+    fn description(&self) -> Option<String> {
+        Some("Reads ODV ASCII files from specified glob paths.".to_string())
+    }
+
+    fn arguments(&self) -> Option<Vec<Field>> {
         Some(vec![Field::new(
             "glob_paths",
             DataType::List(Arc::new(Field::new("glob_path", DataType::Utf8, false))),
@@ -58,24 +59,22 @@ impl BeaconTableFunctionImpl for ReadZarrFunc {
     }
 }
 
-impl TableFunctionImpl for ReadZarrFunc {
+impl TableFunctionImpl for ReadOdvAsciiFunc {
     fn call(
         &self,
-        args: &[datafusion::prelude::Expr],
-    ) -> datafusion::error::Result<std::sync::Arc<dyn datafusion::catalog::TableProvider>> {
-        let glob_paths = crate::file_formats::parse_glob_paths_arg(args, "read_zarr")?;
+        args: &[Expr],
+    ) -> datafusion::error::Result<Arc<dyn datafusion::catalog::TableProvider>> {
+        let glob_paths = beacon_common::table_function::parse_glob_paths_arg(args, "read_odv_ascii")?;
 
-        tracing::debug!("read_zarr glob paths: {:?}", glob_paths);
+        tracing::debug!("read_odv_ascii glob paths: {:?}", glob_paths);
 
         let mut listing_urls = vec![];
         for path in &glob_paths {
-            tracing::debug!("read_zarr processing path: {}", path);
+            tracing::debug!("read_odv_ascii processing path: {}", path);
             listing_urls.push(parse_listing_table_url(&self.data_object_store_url, path)?);
         }
 
-        // Predicate pushdown is handled automatically by the shared engine, so
-        // no manual statistics/column selection is needed.
-        let file_format = ZarrFormat::default();
+        let file_format = OdvFormat::new();
         let super_listing_table = tokio::task::block_in_place(|| {
             self.runtime_handle.block_on(async move {
                 SuperListingTable::new(
