+++ beacon-file-formats/beacon-arrow-parquet/src/datafusion/table_function.rs
@@ -2,21 +2,21 @@ use std::sync::Arc;
 
 use arrow::datatypes::{DataType, Field};
 use beacon_common::{listing_url::parse_listing_table_url, super_table::SuperListingTable};
-use beacon_arrow_ipc::datafusion::ArrowFormat;
+use crate::datafusion::ParquetFormat;
 use datafusion::{
     catalog::TableFunctionImpl, execution::object_store::ObjectStoreUrl, prelude::SessionContext,
 };
 
-use crate::file_formats::BeaconTableFunctionImpl;
+use beacon_common::table_function::BeaconTableFunctionImpl;
 
-pub struct ReadArrowFunc {
+pub struct ReadParquetFunc {
     // Session Reference
     runtime_handle: tokio::runtime::Handle,
     session_ctx: Arc<SessionContext>,
     data_object_store_url: ObjectStoreUrl,
 }
 
-impl ReadArrowFunc {
+impl ReadParquetFunc {
     pub fn new(
         runtime_handle: tokio::runtime::Handle,
         session_ctx: Arc<SessionContext>,
@@ -30,23 +30,23 @@ impl ReadArrowFunc {
     }
 }
 
-impl std::fmt::Debug for ReadArrowFunc {
+impl std::fmt::Debug for ReadParquetFunc {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "ReadArrowFunc")
+        write!(f, "ReadParquetFunc")
     }
 }
 
-impl BeaconTableFunctionImpl for ReadArrowFunc {
+impl BeaconTableFunctionImpl for ReadParquetFunc {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
 
-    fn name(&self) -> String {
-        "read_arrow".to_string()
+    fn description(&self) -> Option<String> {
+        Some("Reads Parquet files from specified glob paths.".to_string())
     }
 
-    fn description(&self) -> Option<String> {
-        Some("Reads Arrow files from specified glob paths.".to_string())
+    fn name(&self) -> String {
+        "read_parquet".to_string()
     }
 
     fn arguments(&self) -> Option<Vec<arrow::datatypes::Field>> {
@@ -58,22 +58,22 @@ impl BeaconTableFunctionImpl for ReadArrowFunc {
     }
 }
 
-impl TableFunctionImpl for ReadArrowFunc {
+impl TableFunctionImpl for ReadParquetFunc {
     fn call(
         &self,
         args: &[datafusion::prelude::Expr],
     ) -> datafusion::error::Result<std::sync::Arc<dyn datafusion::catalog::TableProvider>> {
-        let glob_paths = crate::file_formats::parse_glob_paths_arg(args, "read_arrow")?;
+        let glob_paths = beacon_common::table_function::parse_glob_paths_arg(args, "read_parquet")?;
 
-        tracing::debug!("read_arrow glob paths: {:?}", glob_paths);
+        tracing::debug!("read_parquet glob paths: {:?}", glob_paths);
 
         let mut listing_urls = vec![];
         for path in &glob_paths {
-            tracing::debug!("read_arrow processing path: {}", path);
+            tracing::debug!("read_parquet processing path: {}", path);
             listing_urls.push(parse_listing_table_url(&self.data_object_store_url, path)?);
         }
 
-        let file_format = ArrowFormat::default();
+        let file_format = ParquetFormat::default();
         let super_listing_table = tokio::task::block_in_place(|| {
             self.runtime_handle.block_on(async move {
                 SuperListingTable::new(
